refactor(ecology): split feature substrate rules into files
