+++ mods/mod-swooper-maps/src/domain/ecology/ops/compute-feature-substrate/rules/river-adjacency-mask.ts
@@ -1,0 +1,42 @@
+import { clampInt } from "@swooper/mapgen-core/lib/math";
+
+export function computeRiverAdjacencyMask(args: {
+  width: number;
+  height: number;
+  riverClass: Uint8Array;
+  radius: number;
+}): Uint8Array {
+  const width = args.width | 0;
+  const height = args.height | 0;
+  const radius = clampInt(args.radius | 0, 0, Math.max(width, height));
+  const size = Math.max(0, width * height);
+
+  const mask = new Uint8Array(size);
+  if (radius <= 0) {
+    for (let i = 0; i < size; i++) mask[i] = args.riverClass[i] ? 1 : 0;
+    return mask;
+  }
+
+  for (let y = 0; y < height; y++) {
+    const y0 = Math.max(0, y - radius);
+    const y1 = Math.min(height - 1, y + radius);
+    for (let x = 0; x < width; x++) {
+      const x0 = Math.max(0, x - radius);
+      const x1 = Math.min(width - 1, x + radius);
+      let adjacent = 0;
+      for (let ny = y0; ny <= y1 && !adjacent; ny++) {
+        const row = ny * width;
+        for (let nx = x0; nx <= x1; nx++) {
+          if (args.riverClass[row + nx] !== 0) {
+            adjacent = 1;
+            break;
+          }
+        }
+      }
+      mask[y * width + x] = adjacent;
+    }
+  }
+
+  return mask;
+}
+
